requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
boto3==1.37.37
botocore==1.37.37
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chromadb
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.14
exceptiongroup==1.1.2
fastapi==0.100.1
faiss-cpu
frozenlist==1.4.0
google-auth==2.22.0
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
kubernetes==27.2.0
langchain
langsmith==0.0.26
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
s3transfer==0.11.5
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tiktoken
unstructured
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websocket-client==1.6.2
websockets==11.0.3
yarl==1.9.2
